Add files via upload
This interactive dashboard investigates whether recession can impact Warner Bros, or the movies industry's profitability in general.

After finalizing the main focus of the story, we created an excel sheet solely showing movies distributed by Warner Bros & given the recession sheet as well, a VLOOKUP of the release date and observation date was done, so they could match as either expansion or recession. 

For Tableau functionalities, we created a calculation field using window_sum is used to find the average and display the reference line. From there, we created a second calculation field for the coloring of the graph by grabbing every value for both recession and expansion that is classified above the average line which was done in the previous calculation. This is to be shown as the color green or above average, and everything else is classified as below average and showcased in red. This feature allows stakeholders to filter the data according to their needs, thus enhancing the interactivity of the dashboard. Lastly, for the DVD sales and US sales quadrant scatter plot, we firstly created a set for top 20 movies by DVD sales, then used calculated fields to set up conditions and segment the movies to each quadrant. 

By using parameters and filters, we can observe the effect of recessions on Warner Bros production. For the first period of recession, both the US and worldwide revenue of Warner Bros went below the average. However, during the 2008 recession, Warner Bros revenue still went above average. During the expansionary period (2004-2007), Warner Bros’ gross revenue (both US and Worldwide) were above average.